pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "configgraphviz"
version = "0.1.0"
authors = [{ name = "Richard CHAN", email = "richard.mychan@gmail.com" }]
description = "Generate Graphviz DOT visualizations from INI, YAML, and TOML configuration files."
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Typing :: Typed",
]
keywords = [
    "config",
    "configuration",
    "graphviz",
    "dot",
    "visualize",
    "ini",
    "yaml",
    "toml",
    "parser",
    "diagram",
]
dependencies = [
    "PyYAML>=5.1",
    # Make tomli conditional only if needed before Python 3.11
    'tomli>=1.1.0; python_version < "3.11"',
]

[project.urls]
Homepage = "https://github.com/manyan-chan/ConfigGraphViz"
Repository = "https://github.com/manyan-chan/ConfigGraphViz"

[project.optional-dependencies]
test = ["pytest>=7.0", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["configgraphviz*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
